Major update: Use of DEPHY-SCM case reading + Only compatible with DEPHY format. Works with ARMCU

+++ main/install_SFX_cases.py
@@ -10,10 +10,19 @@
 # eg install_SFX_cases config/config_arp631_CMI6.py
 #
 
-import sys, os
+import os
 
 REP_EMS = os.getenv('REP_EMS')
 REP_MUSC = os.getenv('REP_MUSC')
+
+import sys
+
+sys.path.append('{0}/UTIL/install')
+
+import argparse
+
+import EMS_cases as CC
+import install_MUSC
 
 import configmain
 import configsim
@@ -22,23 +31,20 @@ if not(configsim.lsurfex):
   print 'SURFEX is disabled in configsim.py'
   sys.exit()
 
-argv = sys.argv
-try:
-    config_file = argv[1]
-    print 'your configuration file is', config_file
-except:
-    print "ERROR: you must give a configuration file, eg install_SFX_cases config/config_arp631_CMI6.py"
-    sys.exit()
+# Definition of arguments
+parser = argparse.ArgumentParser()
+parser.add_argument("-case", help="case", type=str, required=True)
+parser.add_argument("-subcase", help="subcase (default: REF)", type=str, default="REF")
+parser.add_argument("-config", help="configuration file", type=str, required=True)
 
-# In case a case/subcase is provided in argument
-lasked = False
-try:
-    case = argv[2]
-    SUB = argv[3]
-    lasked = True
-except:
-    pass
+# Getting arguments
+args = parser.parse_args()
+case = args.case
+subcase = args.subcase
+config_file = args.config
+print 'your configuration file is', config_file
 
+# Importing config file
 try:
     os.remove('config.py')
 except:
@@ -48,9 +54,8 @@ os.symlink(config_file,'./config.py')
 
 import config as configmod
 
-import EMS_cases as CC
-import install_MUSC
 
+# Installing SURFEX for case/subcase
 repout = REP_MUSC + '/SURFEX/'
 
 config = configmod.config 
@@ -59,31 +64,21 @@ PGD = configmod.PGD
 PREP = configmod.PREP 
 namSFXref = configmod.namSFXref 
 
-if configsim.allcases:
-  cases = CC.cases
-else:
-  cases = configsim.cases
-
-if lasked:
-  cases = [case,]
-
 loverwrite = configmain.loverwrite
 lupdate = configmain.lupdate
 
-subcases = CC.subcases
 data_input = CC.data_input
 
-for case in cases:
-  print case    
-  if subcases.has_key(case):
-    SS = subcases[case]
-    if lasked: SS = [SUB,]      
-    for subcase in SS:	
-      print subcase
-      install_MUSC.install_SFX(configsim.model,case,data_input[case][subcase],repout,cycle,PGD,PREP,config,namSFXref,subcase=subcase,loverwrite=loverwrite,lupdate=lupdate)
-  else:
-    install_MUSC.install_SFX(configsim.model,case,data_input[case],repout,cycle,PGD,PREP,config,namSFXref,loverwrite=loverwrite,lupdate=lupdate)
-      
+if not(case in CC.cases):
+    print 'ERROR: case {0} is not known'.format(case)
+    print 'ERROR: known cases:', CC.cases
+    sys.exit()
+
+if not(subcase in CC.subcases[case]):
+    print 'ERROR: subcase {0} is not known for case {1}'.format(subcase,case)
+    print 'ERROR: known subcases for case {0}:'.format(case), CC.subcases[case]
+
+install_MUSC.install_SFX(configsim.model,case,subcase,data_input[case][subcase],repout,cycle,PGD,PREP,config,namSFXref,loverwrite=loverwrite,lupdate=lupdate)
 
 os.remove('config.py')      
 os.remove('config.pyc')
